+++ modele/admin_sup_peintre.php
@@ -8,19 +8,26 @@ $sup_peinture= $requete->fetch(PDO::FETCH_OBJ);
 
 $chemin = "vue/img/peinture/";
 
-if(is_dir($chemin)) {
-    // on récupère le nom des fichiers à effacer (on ne peut effacer qu'un dossier vide)
-    $liste_fichier = scandir($chemin);
-    unlink("$chemin/$sup_peinture->imgsrc");
-}
+
+if($_SESSION['droit'] != 1) {
+    $affiche_demo = true;
+    $affiche_success = false;
+    header("Location: ./");
+}else {
+    if (is_dir($chemin)) {
+        // on récupère le nom des fichiers à effacer (on ne peut effacer qu'un dossier vide)
+        $liste_fichier = scandir($chemin);
+        unlink("$chemin/$sup_peinture->imgsrc");
+    }
 
 
 // suppression de l'image (ligne correspondante) dans la base de données
-try {
-    $eventsupprime = $dbh->exec("DELETE FROM peinture WHERE id= $sup;");
-    header("Location: ?peinture");
-}catch (Exception $e) {
-    echo "Erreur :".$e->getMessage();
-    die();
-}
+    try {
+        $eventsupprime = $dbh->exec("DELETE FROM peinture WHERE id= $sup;");
+        header("Location: ?peinture");
+    } catch (Exception $e) {
+        echo "Erreur :" . $e->getMessage();
+        die();
+    }
 
+}
